refactor(lib): centralize GitHub API headers into github-api-helpers
The request-header builder was copy-pasted as `githubHeaders()` /
`searchHeaders()` / inline objects across 5 github-* modules, and the
SEARCH_PER_PAGE/SEARCH_MAX_PAGES constants in 2 of them. Extracted to
lib/github-api-helpers.ts (getGithubApiHeaders + GITHUB_SEARCH_* constants);
each consumer imports them, aliasing to its prior local names so call sites are
unchanged.

Behavior-preserving (identical headers/values). Verified: github test suites
green (86 tests / 5 suites), tsc --noEmit + eslint clean.
